backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
black==25.1.0
celery==5.5.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-yasg==1.21.10
flake8==7.2.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.1.0
kombu==5.5.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
